// src/app/core/models/doctor.model.ts
export interface Doctor {

  id: number;

  fullName: string;

  email: string;

  specialization: string;

  qualification: string;

  experience: number;

  consultationFee: number;

  availableFrom: string;

  availableTo: string;

  departmentId: number;

  departmentName: string;

  userId: number;
}